% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_kirby21_data.R
\name{copy_kirby21_data}
\alias{copy_kirby21_data}
\title{Copy Kirby21 Data to an output directory}
\usage{
copy_kirby21_data(outdir, ...)
}
\arguments{
\item{outdir}{Output directory for data}

\item{...}{Arguments to pass to \code{\link{get_image_filenames}}}
}
\value{
Logical if files are copied
}
\description{
Copies files from Kirby21 Package to an output directory
}

