% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image_filenames.R
\name{get_image_filenames_df}
\alias{get_image_filenames_df}
\alias{get_image_filenames_matrix}
\title{Get Image Filenames in a matrix}
\usage{
get_image_filenames_df(...)

get_image_filenames_matrix(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{get_image_filenames}}}
}
\description{
Return the filenames for the images
}

