% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kirby21_demog.R
\name{kirby21_demog}
\alias{kirby21_demog}
\title{Kirby 21 Demographics}
\format{A data frame with 21 rows and 3 columns.}
\source{
\url{https://www.nitrc.org/frs/?group_id=313}
}
\description{
A dataset containing demographic information for kirby21
data sets
}

