% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subject_id_to_visit_id.R
\name{subject_id_to_visit_id}
\alias{subject_id_to_visit_id}
\title{Kirby 21 Subject Identifiers to NITRC Visit Identifiers}
\format{A data frame with 42 rows and 4 columns}
\source{
\url{https://www.nitrc.org/frs/?group_id=313}
}
\description{
A dataset containing the mapping from the Subject IDs from the 
Kirby demographics to the KKI2009 idenifiers on NITRC
}

